Refactoring: move Bug to parent package
