+++ java/code/src/com/redhat/rhn/domain/errata/Bug.hbm.xml
@@ -3,7 +3,7 @@
 PUBLIC "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
 "classpath://org/hibernate/hibernate-mapping-3.0.dtd">
 <hibernate-mapping>
-    <class name="com.redhat.rhn.domain.errata.impl.Bug"
+    <class name="com.redhat.rhn.domain.errata.Bug"
         table="rhnErrataBuglist" >
         <composite-id unsaved-value="any">
             <key-many-to-one name="errata"
