+++ java/code/src/com/redhat/rhn/domain/errata/Errata.hbm.xml
@@ -49,7 +49,7 @@ PUBLIC "-//Hibernate/Hibernate Mapping DTD 3.0//EN"
             cascade="all">
             <key column="errata_id"/>
             <one-to-many
-                class="com.redhat.rhn.domain.errata.impl.Bug"/>
+                class="com.redhat.rhn.domain.errata.Bug"/>
         </set>
         <set name="cves" lazy="true" cascade="save-update" table="rhnErrataCve">
                 <key column="errata_id"/>
