+++ java/code/src/com/redhat/rhn/domain/errata/test/ErrataFactoryTest.java
@@ -25,7 +25,7 @@ import com.redhat.rhn.domain.errata.Errata;
 import com.redhat.rhn.domain.errata.ErrataFactory;
 import com.redhat.rhn.domain.errata.ErrataFile;
 import com.redhat.rhn.domain.errata.Severity;
-import com.redhat.rhn.domain.errata.impl.Bug;
+import com.redhat.rhn.domain.errata.Bug;
 import com.redhat.rhn.domain.errata.impl.ClonedErrata;
 import com.redhat.rhn.domain.org.Org;
 import com.redhat.rhn.domain.org.OrgFactory;
