+++ java/code/src/com/redhat/rhn/frontend/xmlrpc/errata/ErrataHandler.java
@@ -39,7 +39,7 @@ import com.redhat.rhn.domain.errata.CveFactory;
 import com.redhat.rhn.domain.errata.Errata;
 import com.redhat.rhn.domain.errata.ErrataFactory;
 import com.redhat.rhn.domain.errata.Severity;
-import com.redhat.rhn.domain.errata.impl.Bug;
+import com.redhat.rhn.domain.errata.Bug;
 import com.redhat.rhn.domain.errata.impl.Keyword;
 import com.redhat.rhn.domain.org.Org;
 import com.redhat.rhn.domain.rhnpackage.Package;
