+++ java/code/src/com/redhat/rhn/manager/errata/ErrataManager.java
@@ -45,7 +45,7 @@ import com.redhat.rhn.domain.errata.Errata;
 import com.redhat.rhn.domain.errata.ErrataFactory;
 import com.redhat.rhn.domain.errata.ErrataFile;
 import com.redhat.rhn.domain.errata.Severity;
-import com.redhat.rhn.domain.errata.impl.Bug;
+import com.redhat.rhn.domain.errata.Bug;
 import com.redhat.rhn.domain.errata.impl.ClonedErrata;
 import com.redhat.rhn.domain.image.ImageInfo;
 import com.redhat.rhn.domain.org.Org;
